+++ controller/SchedController.cs
@@ -1,5 +1,5 @@
 using CriteriosDominio.Dominio.Modelos.Entidades;
-using Infrastructure.src.interfaces;
+using CriteriosDominio.Dominio.interfaces;
 using Microsoft.AspNetCore.Mvc;
 
 namespace CriteriosDeProgramacion.controller
@@ -44,8 +44,16 @@ namespace CriteriosDeProgramacion.controller
         [HttpDelete("{id}")]
         public IActionResult DeleteSched(int id)
         {
-            _schedRepository.DeleteSched(id);
-            return Ok("Sched eliminado correctamente");
+
+            if (_schedRepository.GetSchedById(id) == null)
+            {
+                return BadRequest("No se encontro el sched");
+            }
+            else
+            {
+                _schedRepository.DeleteSched(id);
+                return Ok("Sched eliminado correctamente");
+            }
         }
 
     }
