+++ src/Infrastructure/src/repository/SchedRepository.cs
@@ -1,6 +1,6 @@
 using Infrastructure.contexto;
 using CriteriosDominio.Dominio.Modelos.Entidades;
-using Infrastructure.src.interfaces;
+using CriteriosDominio.Dominio.interfaces;
 using CriteriosDominio.Dominio.Servicios;
 using CriteriosDominio.Dominio.Helpers;
 
@@ -82,9 +82,13 @@ namespace Infrastructure.src.repository
             using (var context = new AppDbContext())
             {
                 var sched = context.Sched.FirstOrDefault(x => x.SchedId == id);
-                context.Sched.Remove(sched);
-                context.SaveChanges();
+                if (sched != null)
+                {
+                    context.Sched.Remove(sched);
+                    context.SaveChanges();
+                }
             }
         }
+
     }
 }
